XRDDEV-746 Move general purpose ui components to a new directory
