+++ src/proxy-ui-api/frontend/src/components/TokenLoginDialog.vue
@@ -37,7 +37,7 @@
 <script lang="ts">
 import Vue from 'vue';
 import { ValidationProvider, ValidationObserver } from 'vee-validate';
-import SimpleDialog from '@/components/SimpleDialog.vue';
+import SimpleDialog from '@/components/ui/SimpleDialog.vue';
 import * as api from '@/util/api';
 
 export default Vue.extend({
