+++ src/proxy-ui-api/frontend/src/components/ui/ConfirmDialog.vue
@@ -18,7 +18,7 @@
  */
 
 import Vue from 'vue';
-import SimpleDialog from '@/components/SimpleDialog.vue';
+import SimpleDialog from '@/components/ui/SimpleDialog.vue';
 
 export default Vue.extend({
   components: {
