+++ src/proxy-ui-api/frontend/src/views/Certificate.vue
@@ -31,10 +31,10 @@ import Vue from 'vue';
 import * as api from '@/util/api';
 import { mapGetters } from 'vuex';
 import { Permissions } from '@/global';
-import SubViewTitle from '@/components/SubViewTitle.vue';
+import SubViewTitle from '@/components/ui/SubViewTitle.vue';
 import CertificateInfo from '@/components/CertificateInfo.vue';
-import ConfirmDialog from '@/components/ConfirmDialog.vue';
-import LargeButton from '@/components/LargeButton.vue';
+import ConfirmDialog from '@/components/ui/ConfirmDialog.vue';
+import LargeButton from '@/components/ui/LargeButton.vue';
 import CertificateHash from '@/components/CertificateHash.vue';
 
 export default Vue.extend({
