+++ src/proxy-ui-api/frontend/src/views/Service.vue
@@ -189,11 +189,11 @@ import _ from 'lodash';
 import { mapGetters } from 'vuex';
 import { Permissions } from '@/global';
 import * as api from '@/util/api';
-import SubViewTitle from '@/components/SubViewTitle.vue';
+import SubViewTitle from '@/components/ui/SubViewTitle.vue';
 import AccessRightsDialog from '@/components/AccessRightsDialog.vue';
-import ConfirmDialog from '@/components/ConfirmDialog.vue';
-import HelpIcon from '@/components/HelpIcon.vue';
-import LargeButton from '@/components/LargeButton.vue';
+import ConfirmDialog from '@/components/ui/ConfirmDialog.vue';
+import HelpIcon from '@/components/ui/HelpIcon.vue';
+import LargeButton from '@/components/ui/LargeButton.vue';
 import { Service, AccessRightSubject } from '@/types.ts';
 import { isValidWsdlURL } from '@/util/helpers';
 import {
