+++ src/proxy-ui-api/frontend/src/views/ServiceDescriptionDetails.vue
@@ -109,10 +109,10 @@ import { ValidationProvider, ValidationObserver } from 'vee-validate';
 import { mapGetters } from 'vuex';
 import { Permissions } from '@/global';
 import * as api from '@/util/api';
-import SubViewTitle from '@/components/SubViewTitle.vue';
-import ConfirmDialog from '@/components/ConfirmDialog.vue';
+import SubViewTitle from '@/components/ui/SubViewTitle.vue';
+import ConfirmDialog from '@/components/ui/ConfirmDialog.vue';
 import WarningDialog from '@/components/WarningDialog.vue';
-import LargeButton from '@/components/LargeButton.vue';
+import LargeButton from '@/components/ui/LargeButton.vue';
 
 export default Vue.extend({
   components: {
